Split waste schema tests into separate files for clarity This improves modularity and clarity by separating different schema validation concerns.
