+++ src/schemas/waste-hazardous.test.js
@@ -1,81 +1,7 @@
 import { receiveMovementRequestSchema } from './receipt.js'
+import { createTestPayload } from './test-helpers/waste-test-helpers.js'
 
-// Test Constants
-const TEST_CONSTANTS = {
-  SITE_ID: 'site123',
-  VALID_EWC_CODE: '010101',
-  DEFAULT_WASTE_DESCRIPTION: 'Test waste',
-  DEFAULT_PHYSICAL_FORM: 'Solid',
-  DEFAULT_METRIC: 'Tonnes',
-  DEFAULT_AMOUNT: 1,
-  DEFAULT_IS_ESTIMATE: false
-}
-
-// Single flexible payload helper function
-const createTestPayload = (overrides = {}) => {
-  const { wasteItemOverrides, ...rootOverrides } = overrides
-
-  // Build waste item with defaults
-  const defaultWasteItem = {
-    ewcCodes: [TEST_CONSTANTS.VALID_EWC_CODE],
-    wasteDescription: TEST_CONSTANTS.DEFAULT_WASTE_DESCRIPTION,
-    physicalForm: TEST_CONSTANTS.DEFAULT_PHYSICAL_FORM,
-    weight: {
-      metric: TEST_CONSTANTS.DEFAULT_METRIC,
-      amount: TEST_CONSTANTS.DEFAULT_AMOUNT,
-      isEstimate: TEST_CONSTANTS.DEFAULT_IS_ESTIMATE
-    }
-  }
-
-  // Merge waste item overrides
-  const wasteItem = wasteItemOverrides
-    ? { ...defaultWasteItem, ...wasteItemOverrides }
-    : defaultWasteItem
-
-  // Build and return complete payload
-  return {
-    receivingSiteId: TEST_CONSTANTS.SITE_ID,
-    wasteItems: [wasteItem],
-    ...rootOverrides
-  }
-}
-
-describe('Receipt Schema Validation', () => {
-  describe('POPs Indicator Validation', () => {
-    it('should accept valid POPs indicator (true)', () => {
-      const payload = createTestPayload({
-        wasteItemOverrides: { pops: { containsPops: true } }
-      })
-      const result = receiveMovementRequestSchema.validate(payload)
-      expect(result.error).toBeUndefined()
-    })
-
-    it('should accept valid POPs indicator (false)', () => {
-      const payload = createTestPayload({
-        wasteItemOverrides: { pops: { containsPops: false } }
-      })
-      const result = receiveMovementRequestSchema.validate(payload)
-      expect(result.error).toBeUndefined()
-    })
-
-    it('should accept missing POPs section', () => {
-      const payload = createTestPayload()
-      const result = receiveMovementRequestSchema.validate(payload)
-      expect(result.error).toBeUndefined()
-    })
-
-    it('should reject missing containsPops field', () => {
-      const payload = createTestPayload({
-        wasteItemOverrides: { pops: {} } // Empty pops object without containsPops
-      })
-      const result = receiveMovementRequestSchema.validate(payload)
-      expect(result.error).toBeDefined()
-      expect(result.error.message).toContain(
-        'Does the waste contain persistent organic pollutants (POPs)? is required'
-      )
-    })
-  })
-
+describe('Receipt Schema Validation - Hazardous', () => {
   describe('Hazardous Waste Validation', () => {
     // Helper function to validate a payload with hazardous waste data
     const validateHazardous = (hazardous) => {
@@ -380,119 +306,6 @@ describe('Receipt Schema Validation', () => {
     })
   })
 
-  describe('EWC Code Validation', () => {
-    // Helper function to validate a payload with a specific EWC code
-    const validateEwcCode = (ewcCodeArray) => {
-      const payload = createTestPayload({
-        wasteItemOverrides: {
-          ewcCodes: ewcCodeArray,
-          pops: { containsPops: false }
-        }
-      })
-      return receiveMovementRequestSchema.validate(payload)
-    }
-
-    it('should accept valid EWC codes without spaces', () => {
-      // Test with valid EWC codes without spaces
-      const result1 = validateEwcCode(['010101'])
-      const result2 = validateEwcCode(['020101'])
-      const result3 = validateEwcCode(['150101'])
-
-      expect(result1.error).toBeUndefined()
-      expect(result2.error).toBeUndefined()
-      expect(result3.error).toBeUndefined()
-    })
-
-    it('should reject EWC codes with spaces', () => {
-      // Test with valid EWC codes with spaces
-      const result1 = validateEwcCode(['01 01 01'])
-
-      expect(result1.error).toBeDefined()
-      expect(result1.error.message).toContain(
-        '"wasteItems[0].ewcCodes[0]" must be a valid 6-digit numeric code'
-      )
-    })
-
-    it('should reject EWC codes with invalid format', () => {
-      // Test with codes that don't match the 6-digit format
-      const result1 = validateEwcCode(['1234'])
-      const result2 = validateEwcCode(['12345'])
-      const result3 = validateEwcCode(['1234567'])
-      const result4 = validateEwcCode(['ABCDEF'])
-
-      expect(result1.error).toBeDefined()
-      expect(result1.error.message).toContain(
-        'must be a valid 6-digit numeric code'
-      )
-
-      expect(result2.error).toBeDefined()
-      expect(result2.error.message).toContain(
-        'must be a valid 6-digit numeric code'
-      )
-
-      expect(result3.error).toBeDefined()
-      expect(result3.error.message).toContain(
-        'must be a valid 6-digit numeric code'
-      )
-
-      expect(result4.error).toBeDefined()
-      expect(result4.error.message).toContain(
-        'must be a valid 6-digit numeric code'
-      )
-    })
-
-    it('should reject EWC codes not in the official list', () => {
-      // Test with codes that match the format but aren't in the list
-      const result1 = validateEwcCode(['999999'])
-
-      expect(result1.error).toBeDefined()
-      expect(result1.error.message).toContain(
-        'must be a valid EWC code from the official list'
-      )
-    })
-
-    it('should reject EWC codes array with more than 5 items', () => {
-      const result = validateEwcCode([
-        '999999',
-        '999999',
-        '999999',
-        '999999',
-        '999999',
-        '999999'
-      ])
-
-      expect(result.error).toBeDefined()
-      expect(result.error.message).toContain(
-        '"wasteItems[0].ewcCodes[0]" must be a valid EWC code from the official list'
-      )
-    })
-
-    it('should require the EWC code field', () => {
-      // Test with missing EWC code - need to build manually as ewcCodes is required
-      const payload = {
-        receivingSiteId: TEST_CONSTANTS.SITE_ID,
-        wasteItems: [
-          {
-            wasteDescription: TEST_CONSTANTS.DEFAULT_WASTE_DESCRIPTION,
-            physicalForm: TEST_CONSTANTS.DEFAULT_PHYSICAL_FORM,
-            weight: {
-              metric: TEST_CONSTANTS.DEFAULT_METRIC,
-              amount: TEST_CONSTANTS.DEFAULT_AMOUNT,
-              isEstimate: TEST_CONSTANTS.DEFAULT_IS_ESTIMATE
-            }
-          }
-        ]
-      }
-
-      const result = receiveMovementRequestSchema.validate(payload)
-
-      expect(result.error).toBeDefined()
-      expect(result.error.message).toContain(
-        '"wasteItems[0].ewcCodes" is required'
-      )
-    })
-  })
-
   describe('Chemical/Biological Concentration Validation', () => {
     // Helper function to validate a payload with hazardous waste and components
     const validateHazardousWithComponents = (containsHazardous, components) => {
@@ -680,51 +493,6 @@ describe('Receipt Schema Validation', () => {
           'Chemical or Biological concentration cannot be negative'
         )
       })
-    })
-  })
-
-  describe('Physical Form Validation', () => {
-    it('should accept valid physical form', () => {
-      const payload = createTestPayload({
-        wasteItemOverrides: { physicalForm: 'Solid' }
-      })
-      const result = receiveMovementRequestSchema.validate(payload)
-      expect(result.error).toBeUndefined()
-    })
-
-    it('should reject invalid physical form', () => {
-      const payload = createTestPayload({
-        wasteItemOverrides: { physicalForm: 'Invalid' }
-      })
-      const result = receiveMovementRequestSchema.validate(payload)
-      expect(result.error).toBeDefined()
-      expect(result.error.message).toContain(
-        '"wasteItems[0].physicalForm" must be one of [Gas, Liquid, Solid, Powder, Sludge, Mixed]'
-      )
-    })
-
-    it('should reject empty physical form', () => {
-      // Need to build manually as physicalForm is required in defaults
-      const payload = {
-        receivingSiteId: TEST_CONSTANTS.SITE_ID,
-        wasteItems: [
-          {
-            ewcCodes: [TEST_CONSTANTS.VALID_EWC_CODE],
-            wasteDescription: TEST_CONSTANTS.DEFAULT_WASTE_DESCRIPTION,
-            // physicalForm missing,
-            weight: {
-              metric: TEST_CONSTANTS.DEFAULT_METRIC,
-              amount: TEST_CONSTANTS.DEFAULT_AMOUNT,
-              isEstimate: TEST_CONSTANTS.DEFAULT_IS_ESTIMATE
-            }
-          }
-        ]
-      }
-      const result = receiveMovementRequestSchema.validate(payload)
-      expect(result.error).toBeDefined()
-      expect(result.error.message).toContain(
-        '"wasteItems[0].physicalForm" is required'
-      )
     })
   })
