+++ src/schemas/waste-pops.test.js
@@ -1,0 +1,39 @@
+import { receiveMovementRequestSchema } from './receipt.js'
+import { createTestPayload } from './test-helpers/waste-test-helpers.js'
+
+describe('Receipt Schema Validation - POPs', () => {
+  describe('POPs Indicator Validation', () => {
+    it('should accept valid POPs indicator (true)', () => {
+      const payload = createTestPayload({
+        wasteItemOverrides: { pops: { containsPops: true } }
+      })
+      const result = receiveMovementRequestSchema.validate(payload)
+      expect(result.error).toBeUndefined()
+    })
+
+    it('should accept valid POPs indicator (false)', () => {
+      const payload = createTestPayload({
+        wasteItemOverrides: { pops: { containsPops: false } }
+      })
+      const result = receiveMovementRequestSchema.validate(payload)
+      expect(result.error).toBeUndefined()
+    })
+
+    it('should accept missing POPs section', () => {
+      const payload = createTestPayload()
+      const result = receiveMovementRequestSchema.validate(payload)
+      expect(result.error).toBeUndefined()
+    })
+
+    it('should reject missing containsPops field', () => {
+      const payload = createTestPayload({
+        wasteItemOverrides: { pops: {} } // Empty pops object without containsPops
+      })
+      const result = receiveMovementRequestSchema.validate(payload)
+      expect(result.error).toBeDefined()
+      expect(result.error.message).toContain(
+        'Does the waste contain persistent organic pollutants (POPs)? is required'
+      )
+    })
+  })
+})
